+++ CHANGELOG.md
@@ -7,6 +7,12 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.9.4 - Unreleased
 
+
+### Added
+
+* Added `maxbytes` parameter to `hash_file` to allow for only hashing a prefix.
+
+
 ### Changed
 
 * Made import time 13x faster (was 109680, is now 8120) by using lazy
@@ -14,6 +20,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 * Removed import dependency on six. There is still a runtime dependency, but we
   are moving away from six. This is a first step to deprecating Python2 support
+
+* Changed default of "blocksize" in `hash_file` to `2 ** 20` based on benchmarks.
 
 
 ## Version 0.9.3 - Released 2020-10-24